requirements.txt
flask
flask_sqlalchemy
flask_marshmallow
marshmallow_sqlalchemy
psycopg2-binary
flask-migrate
gunicorn==21.2.0